+++ packages/types/src/test/index.ts
@@ -17,7 +17,8 @@ import {
   ICurrency,
   IPromise,
   IPromiseConstructor,
-  IStorage
+  IStorage,
+  IMoney
 } from '../main';
 
 // IStringMap
@@ -177,4 +178,35 @@ const brokenStorage: IStorage = {
   },
   size: () => 42,
   reset: () => {}
+};
+
+// IMoney
+
+const brokenMoney: IMoney = {
+  value: 123,
+  toString() { // $ExpectError
+    return 123;
+  }
+};
+
+const brokenMoney2: IMoney = { // $ExpectError
+  currency: 'bar',
+};
+
+const brokenMoney3: IMoney = {
+  value: 'foo', // $ExpectError
+  currency: 123, // $ExpectError
+};
+
+const brokenMoney4: IMoney = {
+  value: {}, // $ExpectError
+  currency: 'bar',
+};
+
+const money: IMoney = {
+  value: 123,
+  currency: 'RUB',
+  toString() {
+    return this.value.toString();
+  }
 };